Moving to setuptools from poetry since it cannot install the new fairseq version due to a bad dependency definition of omegaconf. Setuptools does not raise the same exception as poetry.
