+++ pyproject.toml
@@ -1,19 +1,8 @@
-[tool.black]
-line-length = 120
-
-[tool.isort]
-profile = "black"
-multi_line_output = 3
-line_length = 120
-
-[tool.pytest.ini_options]
-testpaths = ["tests"]
-
-[tool.poetry]
+[project]
 name = "greynirseq"
 version = "0.4"
 description = "Natural language processing for Icelandic"
-authors = ["Miðeind ehf <tauganet@mideind.is>"]
+authors = [{name = "Miðeind ehf", email="tauganet@mideind.is"}]
 classifiers = [
     "Development Status :: 4 - Beta",
     "Intended Audience :: Developers",
@@ -27,52 +16,64 @@ classifiers = [
     "Environment :: GPU :: NVIDIA CUDA :: 11.1",
     "Programming Language :: Python",
     "Programming Language :: Python :: 3",
-    "Programming Language :: Python :: 3.6",
     "Programming Language :: Python :: 3.7",
     "Programming Language :: Python :: 3.8",
     "Programming Language :: Python :: 3.9",
+    "Programming Language :: Python :: 3.10",
     "Programming Language :: Cython",
     "Topic :: Scientific/Engineering :: Artificial Intelligence",
     "Topic :: Software Development :: Libraries :: Python Modules",
     "Topic :: Utilities",
     "Topic :: Text Processing :: Linguistic",
 ]
-keywords = ["nlp", "pos", "ner", "icelandic"]
-license = "AGPLv3+"
+keywords = ["nlp", "pos", "ner", "icelandic", "translation"]
+license = {text="AGPLv3+"}
+requires-python = ">=3.7.2"
 readme = "README.md"
-homepage = "https://github.com/mideind/GreynirSeq"
-repository = "https://github.com/mideind/GreynirSeq"
-packages = [{ include = "greynirseq", from = "src" }]
-include = ["LICENSE"]
-build = "build.py"
+# homepage = "https://github.com/mideind/GreynirSeq"
+# repository = "https://github.com/mideind/GreynirSeq"
+# include = ["LICENSE"]
+# build = "build.py"
+dependencies = [
+    "fairseq>=0.12.2",
+    "nltk>=3.5",
+    "reynir>=3.5.0",
+    "transformers>=4.3.2",
+    "spacy>=2",
+    "pyjarowinkler>=1.8",
+    "scipy>=1.5",
+    "sentencepiece>=0.1.96",
+    "langid>=1.1.6",
+    "editdistance>=0.6.0"
+]
 
-[tool.poetry.scripts]
+[project.scripts]
 greynirseq = "greynirseq.cli.greynirseq_main:main"
 
-[tool.poetry.dependencies]
-python = "^3.7.2"
-fairseq = "^0.12.2"
-nltk = "^3.5"
-reynir = "^3.5.0"
-transformers = "^4.3.2"
-spacy = "^2"
-pyjarowinkler = "^1.8"
-scipy = "^1.5"
-sentencepiece = "^0.1.96"
-langid = "^1.1.6"
-editdistance = "^0.6.0"
+[project.optional-dependencies]
+dev = [
+    "pydocstyle>=5.0.2",
+    "pytest>=6.0.1",
+    "tensorboard>=2.3.0",
+    "rope>=0.17.0",
+    "black>=22.1.0",
+    "isort>=5.7.0",
+    "pylint>=2.7.2",
+    "flake8>=3.8.4",
+    "mypy>=0.812"
+    ]
 
-[tool.poetry.dev-dependencies]
-pydocstyle = "^5.0.2"
-pytest = "^6.0.1"
-tensorboard = "^2.3.0"
-rope = "^0.17.0"
-black = "^22.1.0"
-isort = "^5.7.0"
-pylint = "^2.7.2"
-flake8 = "^3.8.4"
-mypy = "^0.812"
+[tool.black]
+line-length = 120
+
+[tool.isort]
+profile = "black"
+multi_line_output = 3
+line_length = 120
+
+[tool.pytest.ini_options]
+testpaths = ["tests"]
 
 [build-system]
-requires = ["poetry>=0.12", "cython", "setuptools", "numpy"]
-build-backend = "poetry.masonry.api"
+requires = ["setuptools", "cython", "numpy"]
+build-backend = "setuptools.build_meta"
